# Raise on input/tuple-args length mis-match