--- a/src/cattrs/converters.py
+++ b/src/cattrs/converters.py
@@ -3,7 +3,18 @@ from collections.abc import MutableSet as AbcMutableSet
 from dataclasses import Field
 from enum import Enum
 from functools import lru_cache
-from typing import Any, Callable, Dict, Optional, Tuple, Type, TypeVar, Union
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Generator,
+    Iterable,
+    Optional,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+)
 
 from attr import Attribute
 from attr import has as attrs_has
@@ -84,6 +95,24 @@ def is_optional(typ):
 
 def is_literal_containing_enums(typ):
     return is_literal(typ) and any(isinstance(val, Enum) for val in typ.__args__)
+
+
+def _zip_strict(
+    iterable1: Iterable[T], iterable2: Iterable[V]
+) -> Generator[Tuple[T, V], None, None]:
+    """`zip` but raising if iterables don't finish together."""
+    iterator2 = iter(iterable2)
+    for value1 in iterable1:
+        try:
+            value2 = next(iterator2)
+        except StopIteration:
+            raise ValueError(f"{iterable1} has more items than {iterable2}") from None
+        yield value1, value2
+    try:
+        next(iterator2)
+    except StopIteration:
+        return
+    raise ValueError(f"{iterable2} has more items than {iterable1}")
 
 
 class BaseConverter:
@@ -574,9 +603,17 @@ class BaseConverter:
         else:
             # We're dealing with a heterogenous tuple.
             if self.detailed_validation:
+
+                def iter_types_and_values() -> Tuple[Any, Any]:
+                    try:
+                        yield from _zip_strict(tup_params, obj)
+                    except ValueError as exc_:
+                        exc_.__note__ = f"Structuring {tup} @ index {ix}"
+                        errors.append(exc_)
+
                 errors = []
                 res = []
-                for ix, (t, e) in enumerate(zip(tup_params, obj)):
+                for ix, (t, e) in enumerate(iter_types_and_values()):
                     try:
                         conv = self._structure_func.dispatch(t)
                         res.append(conv(e, t))
@@ -592,7 +629,7 @@ class BaseConverter:
                 return tuple(
                     [
                         self._structure_func.dispatch(t)(e, t)
-                        for t, e in zip(tup_params, obj)
+                        for t, e in _zip_strict(tup_params, obj)
                     ]
                 )