When structuring a heterogeneous tuple, if the input has too few or many elements, this PR now raises an error. The error is a `ValueError` in non-detailed validation.

Two existing tuple-structuring tests had to be fixed (in the zero-element case).

Fixes #57