@Tinche 2022-07-04
Hello, thanks for contributing.

Unfortunately this PR makes the happy path almost twice slower.

```
$ pyperf timeit -g -s 'from cattrs import structure; t = tuple[int, str];v = [1, 2]' 'structure(v, t)'
old: Mean +- std dev: 1.50 us +- 0.04 us
new: Mean +- std dev: 2.94 us +- 0.14 us

# Detailed validation false:
old: Mean +- std dev: 1.39 us +- 0.06 us
new: Mean +- std dev: 1.95 us +- 0.09 us
```

You're overcomplicating it with the `_zip_strict` thing. One length comparison after the transformation should be enough, right?