@EpicWink 2022-07-04
> One length comparison after the transformation should be enough, right?

The one case that that doesn't solve is when `obj` is a un-sized operable with Moreno values than the tuple type. If you're fine with that trade-off, then I'll make the change