@Tinche 2022-07-04
What is `an un-sized operable with Moreno values`?