@EpicWink 2022-07-05
> What is `an un-sized operable with Moreno values`?

Ah, autocorrect. I've updated my previous message