@EpicWink 2022-07-05
Doing the post-check comparing the result length with the input (latest commit) gives benchmark results:
* Detailed validation:
  * Before: 1.96 ± 0.14 μs
  * After: 2.10 ± 0.15 μs
* Not-detailed validation:
  * Before: 1.83 ± 0.09 μs
  * After: 1.98 ± 0.16 μs